pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
